requirements.txt
autopep8==2.0.1
gym-0.26.2
matplotlib==3.8.4
pre-commit==3.0.4
pyglet==2.0.15
pygame==2.5.2
tqdm==4.66.4
torch==2.2.2
ipykernel==6.21.1